Creating a Transform Observation System
Partial Completion within the RunicPillar and Shadow_Rune Game Objects.

The Grid Movement Component will be given an event that will run whenever it has reached a new tile position. This event will be told to tell the transformSubect component to notify all of it's observers of it's new position.

Each Observer can have a Different UpdateObservation() Method. For example, one may send a raycast to all of it's observed components, and if none have direct line of sight (In full Shadow), then it may trigger an event such as unlocking the door.